Refactor font extraction and collection handling
- Simplified the extraction process for TrueType Collections (TTC) by introducing `carve_collection_face` to handle face carving directly.
- Updated `ExtractionError` to encapsulate collection-related errors through `CollectionCarveError`.
- Removed redundant code for face carving from `extract.rs` and centralized it in the `opentype` module.
- Enhanced error handling in `classify_and_unwrap` to properly report collection errors.
- Adjusted tests to validate the new extraction logic and ensure both faces in a collection are accessible.
- Updated documentation to clarify the relationship between font format detection and collection handling.
